Fix ImGui visual source selection crash, relax GLSL typing & streamline startup logging
- Fix ImGui assertion crash (id != window->ID) by enforcing unique ## identifiers and non-blank labels in ParametersPanel launchpad menus and ISF registries.
- Inject GL_ARB_gpu_shader5 and GL_EXT_gpu_shader4 extensions into ISF vertex/fragment shaders to support mixed int/uint operations and legacy shaders (achieving 99.7% compile rate across 327 library shaders).
- Add scoped directory scanning to prevent cross-contamination between sources, filters, and transitions.
- Streamline startup error logging to warnings with debug stack traces.
- Add ISF user library test suite, Gradle build task, architecture docs, and release notes.

+++ src/main/kotlin/llm/slop/liquidlsd/rendering/isf/ISFVisualSource.kt
@@ -81,13 +81,13 @@ class ISFVisualSource(
 
         class BoolInput(val name: String, val param: ModulatableParameter?) : ISFInputBinding() {
             override fun apply(shader: Shader) {
-                shader.setUniform(name, if ((param?.value ?: 0f) > 0.5f) 1.0f else 0.0f)
+                shader.setUniform(name, if ((param?.value ?: 0f) > 0.5f) 1 else 0)
             }
         }
 
         class LongInput(val name: String, val param: ModulatableParameter?) : ISFInputBinding() {
             override fun apply(shader: Shader) {
-                shader.setUniform(name, param?.value ?: 0f)
+                shader.setUniform(name, (param?.value ?: 0f).toInt())
             }
         }
